requirements.txt
fastapi
gunicorn
uvicorn
Jinja2
pytest
requests
pre-commit
python-dotenv
aiofiles
python-multipart
pillow
pytesseract
